requirements.txt
asgiref==3.2.10
Django==3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
json-logging==1.2.0
psycopg2==2.8.5
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
